Extract tiled PNG encoder
